Move gatech strategy config to be configurable
